requirements.txt
asgiref==3.8.1
djangorestframework==3.15.2
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
